Handled double registration case
If a user tries to register with facebook or google but already has an account with that service, just log them in without overwriting any data.
